+++ src/main/java/com/bbdemo/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private long id;
     private String title;
     private String description;
     private String tag;
     private long gmtCreate;
     private long gmtModified;
-    private Integer creator;
+    private long creator;
     private Integer viewCout;
     private Integer likeCout;
     private Integer commentCout;
